refactor: [TEO-376] rename access request download usecase, add missing dto decorator for get access requests
